All features working

+++ controllers/userController.js
@@ -15,7 +15,7 @@ module.exports = {
     // Get a single user by ID
     async getSingleUser(req, res) {
         try {
-            const singleUser = await User.findOne({ _id: req.params.userId }).select('-__v');
+            const singleUser = await User.findOne({ _id: req.params.userId }).select('-__v').populate('thoughts');
 
             if (!singleUser) {
                 return res.status(404).json({ message: 'No user with that ID' });
@@ -30,7 +30,7 @@ module.exports = {
     // Create a new user
     async createUser(req, res) {
         try {
-            const user = await User.create(req.body);
+            const user = (await User.create(req.body));
             res.json(user);
         } catch (err) {
             console.log(err);
@@ -38,7 +38,7 @@ module.exports = {
         }
     },
 
-    // TODO: Update a single user by ID
+    // Update a single user by ID
     async updateUser(req, res) {
         try {
             const updatedUser = await User.findOneAndUpdate(
@@ -57,33 +57,65 @@ module.exports = {
         }
     },
 
-    // Deletes an existing user 
-    // TODO: remove their thoughts when user gets deleted
+    // Deletes an existing user and that user's thoughts
     async deleteUser(req, res) {
-    try {
-        const user = await User.findOneAndDelete({ _id: req.params.userId });
+        try {
+            const singleUser = await User.findOne({ _id: req.params.userId }).select('-__v')
 
-        if (!user) {
-            return res.status(404).json({ message: 'No user exists with that ID' });
+            // Find all thoughts associated with the user and delete them
+            const deletedThoughts = await Thought.deleteMany({ username: singleUser.username });
+
+            if (deletedThoughts.deletedCount === 0) {
+                return res.status(404).json({ message: 'User deleted, but no thoughts found' });
+            }
+
+            const user = await User.findOneAndDelete({ _id: req.params.userId });
+
+            if (!user) {
+                return res.status(404).json({ message: 'No user exists with that ID' });
+            }
+
+            res.json({ message: 'User and their thoughts successfully deleted' });
+        } catch (err) {
+            console.log(err);
+            res.status(500).json(err);
         }
+    },
 
-        // const thoughts = await Thought.findOneAndDelete(
-        //     { students: req.params.userId },
-        //     { $pull: { students: req.params.studentId } },
-        //     { new: true }
-        // );
+    // Create addFriend
+    async addFriend(req, res) {
+        try {
+            // Update the User's friends array
+            await User.findByIdAndUpdate(
+                req.params.userId,
+                { $push: { friends: req.params.friendId } },
+                { runValidators: true, new: true },
+            );
 
-        // if (!thoughts) {
-        //     return res.status(404).json({
-        //         message: 'User deleted, but no thoughts found',
-        //     });
-        // }
+            res.status(200).json('Success');
+        } catch (err) {
+            console.log(err);
+            return res.status(500).json(err);
+        }
+    },
 
-        res.json({ message: 'User successfully deleted' });
-    } catch (err) {
-        console.log(err);
-        res.status(500).json(err);
-    }
-},
+    // TODO: Create removeFriend
+    async removeFriend(req, res) {
+        try {
+            const user = await User.findByIdAndUpdate(
+                req.params.userId,
+                { $pull: { friends: req.params.friendId } },
+                { new: true },
+            );
 
+            if (!user) {
+                return res.status(404).json({ message: 'User not found' });
+            }
+
+            res.status(200).json('Success');
+        } catch (err) {
+            console.log(err);
+            return res.status(500).json(err);
+        }
+    },
 }
